+++ src/brucke.erl
@@ -1,0 +1,32 @@
+%%%
+%%%   Copyright (c) 2018 Klarna Bank AB (publ)
+%%%
+%%%   Licensed under the Apache License, Version 2.0 (the "License");
+%%%   you may not use this file except in compliance with the License.
+%%%   You may obtain a copy of the License at
+%%%
+%%%       http://www.apache.org/licenses/LICENSE-2.0
+%%%
+%%%   Unless required by applicable law or agreed to in writing, software
+%%%   distributed under the License is distributed on an "AS IS" BASIS,
+%%%   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+%%%   See the License for the specific language governing permissions and
+%%%   limitations under the License.
+%%%
+
+%% @doc This is the main module for config validation escript
+-module(brucke).
+
+-export([main/1]).
+
+main([ConfigFile | FilterModuleEbinPaths]) ->
+  ok = code:add_paths(FilterModuleEbinPaths),
+  ok = brucke_config:init(ConfigFile),
+  Discarded = brucke_routes:get_discarded(),
+  Discarded =/= [] andalso error({discarded, Discarded}).
+
+%%%_* Emacs ====================================================================
+%%% Local Variables:
+%%% allout-layout: t
+%%% erlang-indent-level: 2
+%%% End: